// ModernizacionPersonas.Web/ClientApp/src/app/models/IPersonsa.ts
export interface IPersona {
    Codigo: string;
    Correo: any[];
    DigitoVerificacion: string;
    Direccion: any[];
    NumeroDocumento: string;
    FechaNacimiento: string;
    Genero: string;
    PrimerApellido: string;
    PrimerNombre: string;
    SegundoApellido: string;
    SegundoNombre: string;
    Telefono: any[];
    TipoDocumento: string;
    TipoPersona: string;
}